Move unit tests depending on Python language frontend to integration test
