+++ cpg-mcp/src/integrationTest/kotlin/mcp/ApplyConceptsTest.kt
