+++ cpg-mcp/src/integrationTest/kotlin/mcp/CpgLlmAnalyzeToolTest.kt
