+++ cpg-mcp/src/integrationTest/resources/simple_config.py
