+++ cpg-mcp/src/integrationTest/resources/test_example.py
