+++ cpg-mcp/src/test/kotlin/de/fraunhofer/aisec/cpg/mcp/CpgAnalyzeToolTest.kt
@@ -26,74 +26,14 @@
 package de.fraunhofer.aisec.cpg.mcp
 
 import de.fraunhofer.aisec.cpg.mcp.mcpserver.configureServer
-import de.fraunhofer.aisec.cpg.mcp.mcpserver.tools.addCpgAnalyzeTool
-import de.fraunhofer.aisec.cpg.mcp.mcpserver.tools.globalAnalysisResult
-import de.fraunhofer.aisec.cpg.mcp.mcpserver.tools.runCpgAnalyze
-import de.fraunhofer.aisec.cpg.mcp.mcpserver.tools.utils.CpgAnalysisResult
-import de.fraunhofer.aisec.cpg.mcp.mcpserver.tools.utils.CpgAnalyzePayload
 import io.modelcontextprotocol.kotlin.sdk.*
 import io.modelcontextprotocol.kotlin.sdk.server.Server
-import io.modelcontextprotocol.kotlin.sdk.server.ServerOptions
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import kotlin.test.assertIs
-import kotlin.test.assertNotNull
-import kotlinx.coroutines.test.runTest
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.buildJsonObject
-import kotlinx.serialization.json.put
 
 class CpgAnalyzeToolTest {
 
     private lateinit var server: Server
-
-    @Test
-    fun cpgAnalyzeToolIntegrationTest() = runTest {
-        val info = Implementation(name = "test-cpg-server", version = "1.0.0")
-        val options =
-            ServerOptions(
-                capabilities =
-                    ServerCapabilities(tools = ServerCapabilities.Tools(listChanged = true))
-            )
-        server = Server(info, options)
-
-        server.addCpgAnalyzeTool()
-
-        val inputSchema = buildJsonObject {
-            put("content", "def hello():\n    print('Hello World')")
-            put("extension", "py")
-        }
-
-        val request = CallToolRequest(name = "cpg_analyze", arguments = inputSchema)
-
-        val tool = server.tools["cpg_analyze"] ?: error("Tool not registered")
-        val result = tool.handler(request)
-
-        assertNotNull(globalAnalysisResult, "Result should be set after tool execution")
-
-        val resultContent = result.content.firstOrNull()
-        assertIs<TextContent>(resultContent)
-        val resultText = resultContent.text
-        assertNotNull(resultText, "Result content should not be null")
-
-        val analysisResult = Json.decodeFromString<CpgAnalysisResult>(resultText)
-
-        assertEquals(2, analysisResult.functions)
-        assertEquals(1, analysisResult.callExpressions)
-        assertNotNull(analysisResult.nodes)
-    }
-
-    @Test
-    fun cpgAnalyzeToolUnitTest() {
-        val payload =
-            CpgAnalyzePayload(content = "def hello():\n    print('Hello World')", extension = "py")
-        val analysisResult = runCpgAnalyze(payload)
-        assertNotNull(globalAnalysisResult, "Result should be set after tool execution")
-
-        assertEquals(2, analysisResult.functions)
-        assertEquals(1, analysisResult.callExpressions)
-        assertNotNull(analysisResult.nodes)
-    }
 
     @Test
     fun testConfigureServer() {
